tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-fill-mode_forwards_even_alternate.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-fill-mode: forwards ending w/'from'</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-fill-mode-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="If 'animation-fill-mode' is 'forwards',
   'animation-direction' is 'alternate' and iteration count is finite and even,
   the property values in the 'from' or '0%' keyframe are applied after the
   final iteration of the animation">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes slide {
         from {
            margin-left: 110px;
         }
         to {
            margin-left: 210px;
         }
      }
      p + div {
         width: 100px;
         height: 100px;
         background: blue;
         animation-duration: 0.5s;
         animation-delay: 0.5s;
         animation-direction: alternate;
         animation-iteration-count: 2;
         animation-fill-mode: forwards;
      }
      div + div {
         margin-top: -105px;
         margin-left: 105px;
         border: 5px solid black;
         width: 200px;
         height: 100px;
      }
      .test {
         animation-name: slide;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var style = getComputedStyle(div, null);
         var passed = false;
         setTimeout(function() {
            passed = (style.marginLeft == '0px');
         }, 250);
         setTimeout(function() {
            passed = (passed && style.marginLeft == '110px');
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         }, 1750);
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue square below should pause to the left outside the border, then
   jump inside the border and slide right and then left again, stopping at the
   left edge.</p>
   <p>Testing...</p>
   <div></div>
   <div></div>
</body>
</html>
